requirements.txt
pandas
PyYAML
scikit-learn
python-dotenv
loguru